# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AdminLoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  dto.AdminLoginResponse:
    properties:
      token:
        type: string
    type: object
  dto.AdminUserResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      restaurant_id:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  dto.CreateAdminUserRequest:
    properties:
      password:
        type: string
      restaurant_id:
        type: string
      role:
        type: string
      username:
        type: string
    required:
    - password
    - role
    - username
    type: object
...
  dto.CreateRequestRequest:
    properties:
      type:
        $ref: '#/definitions/entity.RequestType'
    required:
    - type
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  dto.RequestResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      status:
        $ref: '#/definitions/entity.RequestStatus'
      table_id:
        type: string
      type:
        $ref: '#/definitions/entity.RequestType'
    type: object
...
  dto.UpdateRequestStatusRequest:
    properties:
      status:
        $ref: '#/definitions/entity.RequestStatus'
    required:
    - status
    type: object
  entity.RequestStatus:
    enum:
    - PENDING
    - IN_PROCESS
    - DONE
    type: string
    x-enum-varnames:
    - Pending
    - InProcess
    - Done
  entity.RequestType:
    enum:
    - CALL_WAITER
    - ASK_BILL
    - ASK_HELP
    type: string
    x-enum-varnames:
    - CallWaiter
    - AskBill
    - AskHelp
host: localhost:8080
info:
  contact: {}
  description: Sistema de llamado de mozos para restaurantes
  title: Waiter API
  version: "1.0"
paths:
  /api/v1/admin/login:
    post:
      consumes:
      - application/json
      description: Valida credenciales y retorna un JWT de administrador con duración
        de 24 horas
      parameters:
      - description: Credenciales de administrador
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AdminLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AdminLoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Login de administrador
      tags:
      - admin
...
  /api/v1/admin/users:
    get:
      description: Superadmin ve todos los usuarios; owner ve solo los employees de
        su restaurante.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.AdminUserResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - AdminToken: []
      summary: Listar usuarios administradores
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Crea un nuevo usuario owner o employee. Superadmin puede crear
        cualquier rol; owner solo puede crear employees de su restaurante.
      parameters:
      - description: Datos del nuevo usuario
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAdminUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AdminUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - AdminToken: []
      summary: Crear usuario administrador
      tags:
      - admin
  /api/v1/admin/users/{id}:
    delete:
      description: Superadmin puede eliminar cualquier usuario; owner solo puede eliminar
        employees de su restaurante.
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - AdminToken: []
      summary: Eliminar usuario administrador
      tags:
      - admin
...
  /api/v1/requests:
    post:
      consumes:
      - application/json
      description: Crea una nueva solicitud de mesa (llamar mozo, pedir cuenta, ayuda).
        Requiere JWT de sesión.
      parameters:
      - description: Bearer <session_token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Tipo de solicitud
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRequestRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RequestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - SessionToken: []
      summary: Crear solicitud
      tags:
      - requests
  /api/v1/requests/{requestId}:
    patch:
      consumes:
      - application/json
      description: Actualiza el estado de una solicitud (marcar como atendida)
      parameters:
      - description: ID de la solicitud
        in: path
        name: requestId
        required: true
        type: string
      - description: Nuevo estado
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRequestStatusRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - AdminToken: []
      summary: Actualizar estado de solicitud
      tags:
      - requests
...
  /api/v1/restaurants/{restaurantId}/requests/active:
    get:
      description: Retorna las solicitudes activas de un restaurante
      parameters:
      - description: ID del restaurante
        in: path
        name: restaurantId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.RequestResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - AdminToken: []
      summary: Obtener solicitudes activas
      tags:
      - requests
...
  /api/v1/tables/{tableId}/status:
    get:
      description: Retorna las solicitudes de una mesa específica
      parameters:
      - description: ID de la mesa
        in: path
        name: tableId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.RequestResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Obtener estado de mesa
      tags:
      - requests
...
securityDefinitions:
  AdminToken:
    description: 'JWT de administrador. Formato: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
  SessionToken:
    description: 'JWT de sesión de cliente (QR). Formato: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"